+++ src/shared/errors/HttpError.ts
